Separate code to make clear what py files do what; add docs.
